Fix creation of user account with empty password
Make it possible to always create a user account without a password
after the "Require a password to use this account" checkbox
in the Anaconda GUI is unchecked.

Previously it was possible to get into a situation when password
validation checks would be active even if the checkbox was unchecked,
preventing user account with empty password from being created in the
Anaconda GUI.

The activation conditions of these checks has been fixed, so it should now
be possible to create a user account with empty password at all times.

Resolves: rhbz#1687315