Don't set default of gpt option at cmdline parsing
Value of config option is always overridden by cmdline parsing result. A
cmdling parsing option with a default value will implicitly override the
corresponding value in a config file.

This patch set set default value of gpt option at cmdline parsing to
'argparse.SUPPRESS', so it is possible to distinguish the case when gpt option
is not passed in cmdline.

Without this fix, it is not possible to override storage.gpt option with a
config file, for example, '/etc/anaconda/profile.d/some.conf', even when gpt
option is not passed at the cmdline.

It is safe to remove default of gpt option at cmdline parsing, since storage.gpt
option defaults to False in the default config.