Support multi-org accounts in the GUI
Make it possible to use Red Hat accounts that are members of multiple
organizations when registering vie the Connect to Red Hat screen of the
Anaconda GUI.

This works by always checking at registration time if a given account
has more than one organization listed. If it does, the registration
attempt is aborted, warning message is shown and a combobox listing the
available organizations displayed.

The user can then pick one of the organizations from the list and try
again. The selected organization id will then be used during the next
authentication attempt, solving the issue.

Resolves: rhbz#1965105