Implement the unified help support
Unify the help support on RHEL and Fedora using new mapping files.
The mappings files are located in the root of the help directory.
For example for RHEL, they are expected to be at:

    /usr/share/anaconda/help/rhel/anaconda-gui.json
    /usr/share/anaconda/help/rhel/anaconda-tui.json

The mapping files contain data about the available help content.
For example:

    {
      "_comment_": [
        "This is a comment",
        "with multiple lines."
      ],
      "_default_": {
        "file": "default-help.xml",
        "anchor": "",
      },
      "installation-summary": {
        "file": "anaconda-help.xml",
        "anchor": "",
      },
      "user-configuration": {
        "file": "anaconda-help.xml",
        "anchor": "creating-a-user-account"
      }
    }

If the mapping file doesn't provide the requested help content,
we will use the default one if specified. Comments are ignored.

Related: rhbz#1995945