Clarify `inst.ks.all` description by using imperative mode
Using imperative mode makes it more clear (by being consistent with
the rest of the manual) that the paragraph is referring to the
`inst.ks.all`, not the default behavior.

Longer version:

Generally, options in manuals etc. are described by imperative mode
("do xz" for option that causes program to do xyz), or by specifically
referring to the option ("with this option..").  Plain informative mode
is used for context or detail.

Using informative mode right away sounds as if the sentence is to
describe *context*, not the given option.  IOW, it's ambiguous whether the
sequential behavior is the default or it's the effect of `inst.ks.all`.