Make organization fetching errors non-fatal
Looks like we can't reliably detect relevant errors encountered
while fetching organization list - while with Satellite we do get
the concrete "UnauthorizedException" with Hosted Candlepin we get
"RestlibException", which could mean many things.

So instead of trying to detect if the error is relevant for fetching
the organization list (eq. wrong username and/or password) let's just
log the raw error message we get (effectively a JSON in string form)
and return an empty organization list.

Still this should not change user visible behavior, as the next task
in line - the username + password registration task, will almost
certainly encounter the same error and will properly report it to the
user.

Resolves: rhbz#2006375