Disable the concurrency check in the error dialog in TUI
The IpmiErrorDialog dialog is run by the error handler. The handler might
be triggered by an error from a different thread. It is possible that we
are already asking for a user input when we decide to show the dialog.
That would violate the concurrency check, so disable it.

Related: rhbz#1807491