Disable the concurrency check in the error dialog in TUI
The IpmiErrorDialog dialog is run by the error handler. The handler might
be triggered by an error from a different thread. It is possible that we
are already asking for a user input when we decide to show the dialog.
That would violate the concurrency check, so disable it.

Resolves: rhbz#1807491
(cherry picked from commit 3900641cb65d35140304e118cf14fbdadb2184ca)