Unify GUI & TUI root spoke completeness conditions
TUI showed Root spoke as completed, even if it was not. This happened because
the default state is "no root password = locked", and the spoke took locked
root as completed too.

That behavior was not consistent with GUI, and has been fixed in RHEL 8.
However, for some reason the fix did not get upstreamed.

With this change:
- GUI and TUI root spokes use identical logic to determine if they are
  mandatory and completed.
- TUI shows both root and user spoke as uncompleted and mandatory on startup.
- TUI shows root spoke as uncompleted but not mandatory if admin user exists.
- The RHEL 8 change is finally ported upstream.

Resolves: rhbz#1999646