Create releases when tags are pushed (#infra)
This runs the existing "make release" autotools target in our existing
"standard" RPM test container. Only minimal changes are needed due to
translation canary dependencies.

The output is a release visible on GitHub that contains the tarball with
anaconda and translations.

Since this is an early version of the workflow, releases are created as
drafts. They have to be manually un-drafted in the GitHub interface
afterwards.