Add support for an RPM source defined by a local path to a repository
Allow to specify an RPM source by a local path to a repository. There are no
real setup tasks to perform, so the path has to be already set up and the
repository ready to be used. This source will be used as a default instead of
the CDROM source to access the repository provided by the stage2 image via the
Dracut's mount point at /run/install/repo.