Enable automatic proposing of Fedora releases in Packit (#infra)
I wasn't able to test this locally which may be expected behavior. So or so, it
will create a PR where we see the proposed change and then we can do more
modification / fixes based on that.

Next part of this change will remove %changelog from the upstream spec file.
Packit should handle that from us. It will generate %changelog from commit headers.