Create a task for mounting a live image
Run a task to mount a live image. Originally, the mount point for the installation
of the image was always /run/install/source and we had to sometimes move mounts
to achieve that. That is too complicated. The task can just return a mount point
that should be used for the installation and that's it.