Fix hashbangs in dracut scripts
Sometimes we relied on bash builtins even when declaring we use sh. Fedora's
initramfs uses bash and we are tightly coupled to that already, anyway, so
let's just be honest about it.

In the one other case, use sh because a hashabang is missign, that is the
more strict variant for the linter, and the file does pass clean as-is.

Overall, this fixes miscellaneous shellcheck warnings related to matching
the shell and used builtins.