Introduce a download path to simplify the image payload code
Run the image download task for all types of the image URL. If the image is
local, the task returns its local location. Otherwise, the image is downloaded
at the specified download location and the task returns that one.

Always run the image removal task, but use the download path to make sure
that we won't remove a local image.