Fix the status of the root configuration screen
If the root account is locked, show the "Root account is disabled" status.
Without the fix, Anaconda shows "Disabled, set password to enable" if there
is `firstboot --reconfig` in a kickstart file.

The workaround for the Initial Setup can be removed. It was implemented in
the commit 98d7b29 for the bug 1507940, but the bug didn't require to add
a special status.

Related: rhbz#2026916