Ignore webui specific parts in the rpm-test
RPM test is testing content of the repository against RPM content, however, that
doesn't work here because JS has more files in the repository than including in
RPM and these in the RPM are gzipped which makes situation even worse.

For now, the best solution to avoid failures of these tests are to ignore the
ui/cockpit part of the repository in these tests. We can try to solve this
better in the future.