Use local anaconda.spec.in file during container build (#infra)
Some of our containers needs dependencies of Anaconda installed in it and to get
these we were downloading anaconda.spec.in from the rhinstaller/anaconda
repository. This works, but it has a big drawbacks like not being able to use
this correctly from a fork for testing or do not test spec file changes in PRs.

Spec file couldn't be read from the repository because podman/docker do allow
only resources under the build directory.

To solve this, the Makefile will copy the anaconda.spec.in file and content of
the required dockerfile directory to temp dir and run the build there.

Side changes:
- we don't need curl in the containers