Simplify the default source selection of the RPM sources
Don't use defaults and fallbacks during the payload setup. Choose the default
source early and only once instead. If the chosen default source is not usable,
a user will have to change it via the user interface or reboot with a different
configuration.

With this change, the installer chooses a default source of the DNF payload
based on data processed in this order:

1. Kickstart file: If there is a source defined in a kickstart file, use it.
2. Boot options or command line options: If there is a source defined in boot
   options or cmdline options, use it. In the future, this may have a higher
   preference than a kickstart file to unify the experience with other boot
   options.
3. Installation image mounted by Dracut: If the installer is booted from an
   image with a repository (for example, the full DVD ISO), use the Dracut's
   mount point at /run/install/repo to access this repository.
4. Anaconda configuration file: If none of the options above is available, use
   the default source defined in the Anaconda configuration file (the closest
   mirror or CDN).

Previously, we were trying to find the installation media and mount it again
using the CDROM source instead of using the Dracut's mount point. That caused
several issues. The installation media could have been different, the new mount
could have failed or Blivet didn't have to see the media because it was marked
as hidden due to other kickstart commands.