Use RPM mount source as the fallback source
The fallback source is a source which is set if the DNF payload doesn't have user
repository set and there is local installation source. There is also a default
source but the local source has a higher priority over default source. The
fallback source is the source detected on the media from which the installation
is started.

The original fallback source was CDRom source but not all the time the CDRom is
really the source of the installation. Instead new RPM Mount source was created
which is just looking if the given directory contains the repository. If yes it
is then used. In the case of fallback source we will look on /run/install/repo
which is directory mounted by our code in Dracut and this directory should
contain the repository (if present on the ISO).