Allow manual releases (#infra)
Add manual starting (workflow_dispatch) to release action.

Input can ingest any string, but it will not work with arbitrary hashes.
The code expects to get a tag that sits on a release commit, so that commit
is automatically skipped. Similarly, it expects that the tag name conforms
to a particular scheme.