Skip /etc/machine-info during live installs (#2036199)
/etc/machine-info should not be installed to the system, for
the same reason we don't install /etc/machine-id: it should be
system-specific, not the same for every install from the same
live image.

This should fix live installs with systemd 250+ not booting due
to the bootloader config snippets being named based on the
KERNEL_INSTALL_MACHINE_ID from the live image /etc/machine-info,
but grub2-mkconfig expecting them to be named according to the
installed system's newly-generated /etc/machine-id.

Signed-off-by: Adam Williamson <awilliam@redhat.com>