Create the LVM devices file, if supported
If Blivet detects that the LVM devices file is suppported, write the file.
Write all PVs to it, so as to keep the previous behavior. Then, copy it to
the new system.

This does not happen if using the file is not supported. Writing the file
with unknown contents (even if empty) would have disastrous consequences,
once the system is updated to a version of LVM that uses this file.

Bumps required Blivet version.

Resolves: rhbz#2011329
Resolves: rhbz#2040302

(cherry picked from commit 8e3223a677a1813ba6c15faead5fb6759b0f80c0)
(cherry picked from commit 6dd7ebbd79103075061c19427737203163d78410)