webui: Introduce template react components for all configuration subpages
Also start implementing the language selection intro menu and the root
password subpage which can be now used as a code reference.

react/prop-types rule was disabled in eslint, as I find that PropTypes
add a lot of code and little help.

Lastly introduce a new test file for the root account subpage, and since
now we have more than one (2!!) test files, let's use the
test/common/run-test wrapper for running the tests in CI.