Always set system purpose from GUI
Previously we tried to detect if system purpose data has changed
in the GUI and only set system purpose to the installation environment
if the data has changed. But this turns out to be error-prone and to not
handle all relevant use-cases, such as re-registration.

Since the DBus based system purpose is configuration (available since
RHEL 9) is essentially free, lets scrap the change detection logic
and just always set system purpose data from the GUI, regardless
from it being the same or not.

This fixes the linked bug and also makes the code simpler and easier
to maintain.

Resolves: rhbz#2015059