build: Remove make as part of run-build-and-arg script
Otherwise re-building webui as part of the container-rpm-test
skips running webpack, resulting in translations js files to not get generated.

Add po-pull target as a dependency to the scratch target.

Translation files should be part of the RPM and there are currently
checks for this in the spec file (only for the webui translations)

Make sure to download translations before building the RPMs, for the
RPM unit checks to pass.