Do not copy resolv.conf to target system at the end of installation
An /etc/resolv.conf symlink is created on target system by
systemd-resolved in postin rpm script.  And if it is not (bug #2018913)
it would be crated during boot of installed system (by systemd-resolved).

The only case where anaconda copying the resolv.conf could have actually
effect would be overwriting /etc/resolv.conf on target system in non-rpm
payload installations (image installations). But even in this case
Anaconda should not interfere into management of /etc/resolv.conf on the
target system. It is a business of either the image creator or a service
on starting installed system (systemd-resolved or NM).