Move the infra-check files check as last check (#infra)
I want to avoid failing the whole check based on this. This has more an
informative character that you should do double review.

Because of that it give sense to have that as the last part of the workflow so
the workflow could strong checks solve first and fail on that.