Comment instead of fail for files check in infra-check (#infra)
The infra-check workflow shouldn't fail when you are changing files which are
not infra specific files. In plenty of situations we need to change Makefile
which is shipped but it's pretty safe to be changed.

So instead of fail just comment on the PR that you should do a double check for
these files.