Remove the support for detection of unsupported hardware
The detection was implemented for RHEL 8 and immediately disabled. There is some
discussion going on about this functionality, but it is very unlikely that we
will be able to enable this code again in its current form.

The removed configuration options can_detect_unsupported_hardware and
can_detect_support_removed were never documented in the RHEL documentation.

Related: rhbh#1645686