UX: clarify meaning of "additional space" checkbox
The phrase "I would like to make additional space available" is ambiguous. It could be interpreted in two ways:
1. I would like to delete existing partitions to reclaim space (the intended meaning)
2. I would like the automatic partitioning to leave some unused space (e.g. because I plan to install something else later)

I wanted to leave some unused space. I checked the box and was surprised when the result occupied my entire hard drive. This change removes the ambiguity.

This is a squashed commit that uses improved wording suggested by @VladimirSlavik