webui: add npm-update workflow and introduce the deploy-keys script
The npm-update workflow automates the process of keeping the
package.json file, in ui/webui directory up to date.
In other words, the dependencies for the webui subproject are not
updated automatically, by the means of creating PRs against this repository.

The PRs will be created by the npm-update workflow, after pushing the changes to a temporary branch
in this repository. For this to be possible, we need a deploy key
[1], that is put into place by the newly introduced deploy-keys.sh
script.

[1] https://docs.github.com/en/developers/overview/managing-deploy-keys