Backport hotfix for the npm issue to anaconda-ci container (#infra)
We don't need that because we don't build the Anaconda in the anaconda-ci
container, however, blivet using our containers for their testing and they are
doing that. So or so, this will also fix the `container-shell` Makefile target
in the similar cases.

The original fix is done by aae34587bc246d31eff2b229f40b1ccc87adb9b5 .