build: clean up webui dist before making scratch builds
If `ui/webui/dist` is present, the `make dist` command will be skipped
for the webui subtree. This is especially problematic, if the
translation files were not present when the `ui/webui/dist` was created
in the first place, as then RPM creation will always fail on missing
translation JS files.