Use cache when fetching organization data from the GUI
Bug 2044258 is caused by GetOrgs() call at registration time
via RetrieveOrganizationsTask triggering MultipleOrganizationsError
- the user is member of more than one organization yet no organization
has been specified.

This fails the registration attempt, which triggers cleanup tasks,
such as un-provisioning from Satellite.

As Satellite is the only known environment where users that are
members of multiple organizations exist, this causes an issue when
the GetOrgs() calling DBus task is called again from the GUI.

The system will no longer be provisioned for Satellite at this point &
the GetOrgs() call will fail - it will go to Hosted Canlepin, which
knowns nothing about this user.

The end result is the behavior described in the bug - the GUI fails to
list organizations for the user to choose from.

To fix this issue, we now always cache the org data list returned by
GetOrgs() when called via RetrieveOrganizationsTask at registration time.
For this the task is called directly, not via its DBus interface and
we make sure the cache is always cleared before calling GetOrgs().

The only case where RetrieveOrganizationsTask is called via it's DBus interface
(RetrieveOrganizationsWithTask) is when GUI needs to list organizations after
handling MultipleOrganizationsError. In this case the existing cached
value will be used when the GetOrgs() call fails due to Satellite being
no longer available.

This way if GUI calls RetrieveOrganizationsWithTask() after handling
the MultipleOrganizationsError, it will get the cached data from the
initial GetOrgs() call, which triggered the MultipleOrganizationsError.

Resolves: rhbz#2044258