Remove --pull-always from default podman build (#infra)
The --pull-always is great but it will always pull the remote repository to get
the base image. It's great to have this for the RHEL/Fedora etc. base images so
you don't work with outdated ones but not that great if you have a container
image based on the something build locally (as anaconda-release container
image).