Fix anaconda-widget crash when loaded on Wayland
The GDK_DISPLAY_XDISPLAY function will return NULL on Wayland. That will than
crash when put into the xkl_engine_get_instance() function. To avoid the crash
let's open the display here directly.

Just for clarification this is not fixing usability on Wayland system, just that
it's not crashing. It still not work correctly.

Resolves: rhbz#1959521

Suggested-by: Ray Strode