Move the po files download to the `make` call
Right now if user will call:

`make && make rpms`

it will fail because webui needs to have translations and `make rpms` won't
create them. To make this more optimized let's download the translations on the
`make` call if they don't present already.

To do that use flag file, which will be created when the files are downloaded
and removed on clean.