Always request localization files during build
After latest changes the translation files are downloaded during the `make`
call, so the localization files should be always available.

Remove this hack which was used in the past to avoid long download times of the
localization from the Zanata system. Now it's just a git clone which is pretty
fast.

This change will also make our current build more robust. The build will fail
in case the translation files are for any reason missing.