Fix Makefile targets using L10N Makefile variables
Move of L10N specific variables to po/Makefile.am break some make targets which
are using these variables in the root Makefile.am. To fix this, move the
variables to configure.ac to have them available everywhere.

This will fix issues with pot file check we have now on the localization
repository.