tests: webui: ensure that installation reaches boot-loader step and fails
This at least ensures that storage configuration step was performed.

Also pass from all steps in the test before clicking 'Begin
installation'.

Skipping steps is not yet guarded in the UI, but since the storage configuration tasks
are only performed after visiting the storage wizard step, we need to go
to that page to let the storage get configured.