Move the validation of system repositories
It is enough to call the load_packages_metadata method of the DNF manager
to load all enabled repositories, but we won't get any error messages that
we could show to users. That's why we load these repositories before that.
Let's separate the validation from the metadata gathering to make it clear.