Fix %onerror scripts not invoked in non-interactive mode
When kickstart's 'cmdline' or 'text' mode is used to
automate the installation if an error is happens due to
incorrect storage configuration or hardware error, it
is not handled properly and any scripts in %onerror
section of the kickstart are not invoked.

This change fixes the issue by handling the exception
properly in those scenarious.