webui: Add Review screen implementation
Add implementation of the Review screen based on latest mockups.

The Review screen shows system configuration just before the
installation is started. All the data on the screen is "live" and
fetched over DBus from the Anaconda backend.

There is also a "selected disks will be overwritten" checkbox that
needs to be checked before the installation can be started.