Run restorecon in chroot when handling home dirs
Running restorecon needs chroot to sysroot so as to have correct absolute
paths. At the point when users are handled, payload has been already
installed, so the SELinux policy is already present.

Note that some parts of this change require Python 3.9+, so this can't be
ported 1:1 to RHEL 8 where the problem has been originally reported.

This is achieved by using the restorecon utility function instead of running
restorecon directly with execWithRedirect.

Related: rhbz#2069305