webui: move language setting of cookies and backend
Previously, we were using React hooks to listen to language changes on the
UI, and updated the backend and cookies on language selection changes.

As a side effect though, the hooks where also called when the LanguageInstallation component
got 'unmounted', in other words, when the user clicked 'Next'.

This caused the 'Installation destination' wizard step to flicker when
loaded.

This commit fixes this problem.