Fix translations in DBus modules (#2071098)
Make sure that LANGUAGE, LC_ALL and LC_MESSAGES are not set in the environment
of DBus modules. Only LANG should be specified. Otherwise, the strings are not
translated based on LANG, but based on one of the other variable.

See: https://docs.python.org/3/library/gettext.html#gettext.find

Resolves: rhbz#2071098