Don't set LC_ALL for live installations
The commit 4d686f1 was created as a fix for the bug 1169019. However, the fix
was rejected and the commit was supposed to be reverted (see the comment 37).

LC_ALL takes precedence over LANG during translations. Anaconda eventually
removes LC_ALL from its environment to get the expected translations, but
it might still affect processes started before that (see the bug 2071098).

Related: rhbz#2071098