Fix keyboard spoke issue if Live system changed keyboard layouts (#2072941)
Fix two issues with the keyboard configuration split on the Fedora Live on
Wayland.

- Anaconda will force user to go into the keyboard spoke when system keyboard is
not default ("en-US")
This issue is happening because Anaconda is not able to correctly read the
system settings. In this case we have a check that user has to adjust keyboard
configuration when system != Anaconda configuration. That is not valid anymore
in case of Wayland where Anaconda is no longer able to control system keyboard
configuration.

- Anaconda will crash in case the Live system has multiple keyboard layouts set
and selected something else than the first one
This bug seems to be there a long time already. As before Anaconda can't read
system keyboard layouts on Wayland so we are looking into the list on a
position get from the system and have shorter list -> we are reading out of the
list.

Both issues are solved by one shot. Move the `can_configure_keyboard()`
check before this check. We can't control system layout so in this
case we shouldn't really force user to check if the configuration is correct.

Resolves: rhbz#2072941