Improve dependabot handling of the infra-check (#infra)
Dependabot will not allow to add your custom postfix which is required by the
infra-check. Because of that we will have a failed check or we have to manually
adjust the commit message.

Let's solve that by allowing `infra:` prefix in the infra-check workflow and
adjust the dependabot configuration to add the prefix.

(cherry picked from commit f4b796cab864708b4633da13488ff2349b6b33e2)