Add missing period when concatenating password error strings (#2075419)
Without this change we have in the bottom line messages like:

"The password fails the dictionary check - it is too simplistic/systematic You
will have to press Done twice to confirm it."

The missing period between sentences could be a big issue for understanding the
sentence in languages like Chinese.

The fix here is not great one because we are not owners of both strings.

The first sentence comes from the libreport library the second one from us. Here
we have two possible solutions. One would be to add translatable
concatenation (what I'm doing here) and second is to add the period into the
`...press Done twice...` message. I don't like to duplicate the second part so
I rather add the new "{}. {}" for translations.

Resolves: rhbz#2075419