Install eslint during container build
Previously, it was installed on first test run, which with containers means
always. Additionally, this installs only eslint, not all of the packages.
The combined savings should be around a minute per test run, at a cost of
ca. +17 MB container size.

This makes the test slightly less correct when run outside the container:
Previously the version to install was taken from the config file. Now, eslint
is automatically installed after invoking it via npx, but it cannot guess
where the version file is, so it installs the latest. However, running the
test without container is unusual, so this should be good enough.