bootloader/base.py: enable resume on arm64
resume=swap_device is only added for x86, thus on arm64 systems resume from
hibernation fails because of that missing kernel command line paramter.

This issue appears on distros that have systemd: systemd uses
systemd-hibernate-resume-generator which sends the parameter of the device to
systemd-hibernate-resume; systemd-hibernate-resume-generator checks the kernel
command line only for the resume device.