Delegate auto-attach handling to RHSM
At the moment both Simple Content Access and "classic" entitlement based
subscription access is supported by Anaconda.

This presents an issue due to auto-attach action handling - for
entitlement subscription to work correctly, we need to register and then
auto-attach.

For SCA to work correctly, we should register and *not* attempt to
auto-attach - the registration action does all that's needed and
auto-attach afterwards is actually not a valid action.

For now this is hot-fixed on Hosted Candlepin side (and likely also on
Satellite side) by ignoring these unnecessary/invalid auto-attach calls
in SCA mode, but longer term the Candlepin team would like to drop this
workaround.

After discussing this issue with the RHSM team, it was decided to handle
this on the RHSM DBus daemon side as the daemon has all the necessary
data (eq. which subscription mode is used, etc.) readily available.

All that is needed on the Anaconda side is to set an option in the RHSM
Register call for RHSM to take over the auto-attach handling & drop all
the auto-attach handling code.

Resolves: rhbz#2083318