Install npm instead of nodejs in our containers (#infra)
The current solution worked, because the npm package is a weak dependency
of the nodejs package. Now it is broken due to some inconsistencies in the
Rawhide compose. It means that npm cannot be installed, but the CI doesn't
fail until it actually tries to use it. Let's install npm instead of nodejs.
The nodejs package will be installed as its required dependency.