Fix growing installation size requirement
Fix issue where the additional space needed for installation (shown
in the hub if the root partition didn't have enough space) would
slightly increase as time went on.

This happened because the size was calculated from the running
file system and included growing log files and other changes done in the
live session.

Delete old method for calculating installation size that uses running
file system and replace it with one that uses the payload file size.
Also modified tests accordingly.