Add infrastructure templating tools
Add a script to scripts/jinja-render for rendering Jinja2 templates.
Add a branch-variables.yml file for storing infrastructure variables of
a current branch.
Add a new target to Makefile.am called refresh-infra for calling this
script on all templates.

In practice if you have an infrastructure file which requires
swapping certain variables depending on the branch, you should
add ".j2" at the end of it's file name and replace all branch dependent
values with "{$ var_name $}" where var_name is a variable defined in
branch-variables.yml. In order to render this template run "make
refresh-infra". Rendered files will have the same path but will be
missing the ".j2" file extension.